+++ Model/Fingering_Lib/VerticalCost.h
@@ -9,7 +9,7 @@ namespace Model
 			VerticalCost() = delete;
 			~VerticalCost() = delete;
 		public:
-			static int Calculate(std::vector<std::pair<int16_t, char>> notes_fingers);
+			static float Calculate(std::vector<std::pair<int16_t, char>> notes_fingers);
 		};
 	}
 }
